feat: add button-group-react package
Based on the creation of the separate Button: https://github.com/nl-design-system/utrecht/commit/1e757386e3cfbaf0efa98a2ef8b9d19536564046#diff-e2954b558f2aa82baff0e30964490d12942e0e251c1aa56c3294de6ec67b7cf5
