+++ .changeset/copy-cat-code.md
@@ -1,0 +1,5 @@
+---
+"@utrecht/button-group-react": major
+---
+
+Add `@utrecht/button-group-react` package.
