+++ packages/component-library-react/package.json
@@ -135,7 +135,7 @@
     "@utrecht/badge-status-css": "workspace:*",
     "@utrecht/blockquote-css": "workspace:*",
     "@utrecht/breadcrumb-nav-css": "workspace:*",
-    "@utrecht/button-group-css": "workspace:*",
+    "@utrecht/button-group-react": "workspace:*",
     "@utrecht/button-link-css": "workspace:*",
     "@utrecht/button-react": "workspace:*",
     "@utrecht/calendar-react": "workspace:*",
