+++ packages/component-library-react/src/css-module/ButtonGroup.tsx
@@ -4,8 +4,4 @@
  * Copyright (c) 2021-2024 Gemeente Utrecht
  */
 
-import css from '@utrecht/button-group-css/dist/index.mjs';
-import { insertStyle } from './insert-style';
-insertStyle(css);
-
-export * from '../ButtonGroup';
+export * from '@utrecht/button-group-react/dist/css';
