+++ packages/components-react/button-group-react/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "env": {
+    "jest/globals": true
+  },
+  "rules": {
+    "react/react-in-jsx-scope": "off"
+  }
+}
