+++ packages/components-react/button-group-react/jest.config.mjs
@@ -1,0 +1,8 @@
+export default {
+  moduleDirectories: ['node_modules', '<rootDir>/'],
+  testEnvironment: 'jest-environment-jsdom',
+  testPathIgnorePatterns: ['/dist/'],
+  moduleNameMapper: {
+    '^@utrecht/(.*)$': '<rootDir>/../$1/src/',
+  },
+};
