+++ packages/components-react/button-group-react/rollup.config.mjs
@@ -1,0 +1,3 @@
+import { createConfig } from '@utrecht/build-utils-react/src/rollup.mjs';
+
+export default createConfig();
