+++ packages/components-react/button-group-react/src/css.tsx
@@ -1,0 +1,12 @@
+/**
+ * @license EUPL-1.2
+ * Copyright (c) 2020-2024 Frameless B.V.
+ * Copyright (c) 2021-2024 Gemeente Utrecht
+ */
+
+import css from '@utrecht/button-group-css/dist/index.mjs';
+import { insertStyle } from './insert-style';
+
+insertStyle(css);
+
+export * from './index';
