+++ packages/components-react/button-group-react/src/insert-style.ts
@@ -1,0 +1,8 @@
+export const insertStyle = function (css: string) {
+  if (typeof document !== 'undefined') {
+    var head = document.head || document.getElementsByTagName('head')[0];
+    var style = document.createElement('style');
+    style.appendChild(document.createTextNode(css));
+    head.appendChild(style);
+  }
+};
